Prefetch repository hash list for local
Doing a listdir on all bucket dirs for the local backend is faster if we
need to backup O(1000) or more files, especially on slow or network
drives, which is the common case. This does delay the start of the
actual backup process by a few seconds. Would be nice if we could do
this async.